feature: add methods to modify the relics DB when inputing texts, api is developing.
